2024-1월 변경사항 적용

+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/price/FloorPriceInfo.java
@@ -47,11 +47,11 @@ public class FloorPriceInfo extends BaseTimeEntity {
 
     private String companyName;
 
-    private int guaranteePrice;
+    private double guaranteePrice;
 
-    private int rent;
+    private double rent;
 
-    private int management;
+    private double management;
 
     private String termStartDate;
